created the model and worked on flask file
I've added the model and done part of the backend. There's still a lit bit to be done you can see from my comments. I also added some reference files to get an idea of how to connect this to the front end. We can deploy on Heroku since it's good for python apps